requirements.txt
fastmcp>=0.4.1
python-dotenv>=1.0.0
notion-client>=2.2.1
todoist-api-python>=2.1.4
